requirements.txt
-i https://pypi.org/simple
altair==5.0.1
altgraph==0.17.3
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cx-freeze==6.15.5
cython==3.0.0
debugpy==1.6.7.post1
frozendict==2.3.8
gitdb==4.0.10
gitpython==3.1.32
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lxml==4.9.3
macholib==1.16.2
markdown-it-py==3.0.0
markupsafe==2.1.3
mdurl==0.1.2
multitasking==0.0.11
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pandas-ta==0.3.14b0
pillow==9.5.0
plost==0.2.5
prettytable==3.8.0
protobuf==4.24.0
pyarrow==12.0.1
pydeck==0.8.0
pyecharts==2.0.3
pygments==2.16.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pympler==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
setuptools==68.0.0
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
streamlit==1.25.0
streamlit-aggrid==0.3.4.post3
streamlit-echarts==0.4.0
streamlit-elements==0.1.0
streamlit-option-menu==0.3.6
streamlit-pills==0.3.0
streamlit-toggle-switch==1.0.2
ta==0.10.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
wcwidth==0.2.6
webencodings==0.5.1
yfinance==0.2.28
zipp==3.16.2
